In packageProduct model we changed the type of total atribut in this case we changed integer by float

+++ model/PackageProduct.js
@@ -23,7 +23,7 @@ const PackageProduct = db.define("packageProduct",{
         type: sequelize.TEXT,
     },
     total: {
-        type: sequelize.INTEGER,
+        type: sequelize.FLOAT,
         allowNull: false
     }
